Version that can merge with master

+++ src/dtgtk/culling.c
@@ -1391,18 +1391,20 @@ static gboolean _thumbs_compute_positions(dt_culling_t *table)
     return TRUE;
   }
 
-  // initialize horizontal and vertical spacing distance between thumbnails with lowest value possible. Will be scaled up later.
+  // initialize horizontal and vertical spacing distance between
+  // thumbnails with lowest value possible. Will be scaled up later.
   const int spacing = 1;
 
-  // reinit size and positions of each thumbnail, remember size from biggest thumbnail, calculate average thumbnail ratio
+  // reinit size and positions of each thumbnail, remember size from
+  // biggest thumbnail, calculate average thumbnail ratio
   int max_thumb_height = 0;
 
   for(GList *l = table->list; l; l = g_list_next(l))
   {
     dt_thumbnail_t *thumb = l->data;
     const float aspect_ratio = thumb->aspect_ratio;
-    thumb->width = (gint)(sqrt(aspect_ratio) * 100);
-    thumb->height = (gint)(1 / sqrt(aspect_ratio) * 100);
+    thumb->width = (gint)(sqrtf(aspect_ratio) * 100);
+    thumb->height = (gint)(1 / sqrtf(aspect_ratio) * 100);
     thumb->x = thumb->y = 0;
 
     max_thumb_height = MAX(max_thumb_height, thumb->height);
@@ -1905,7 +1907,8 @@ void dt_culling_full_redraw(dt_culling_t *table, const gboolean force)
     }
   }
 
-  dt_print(DT_DEBUG_LIGHTTABLE | DT_DEBUG_PERF, "[dt_culling_full_redraw] done in %0.04f sec", dt_get_wtime() - start);
+  dt_print(DT_DEBUG_LIGHTTABLE | DT_DEBUG_PERF,
+           "[dt_culling_full_redraw] done in %0.04f sec", dt_get_wtime() - start);
 
   if(darktable.unmuted & DT_DEBUG_CACHE) dt_mipmap_cache_print();
 }
